Add tentative HBDSCAN* clustering code
Not really runnable on DataHub since UMAP is still chugging along
(it's on run #115/180 right now), but shouldn't be too bad to get
working tomorrow. Main pain point will just be determining majority
superpop. of clusters / seeing if the clusters actually work well
for this data. So, uh, let's wait and find out!